Rename the application models `applmodel_larsimeme` to `applmodel_v1`.
